+++ packages/langchain/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @ai-sdk/langchain
 
+## 2.0.80
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - ai@6.0.75
+
+## 2.0.79
+
+### Patch Changes
+
+- Updated dependencies [471009b]
+  - ai@6.0.74
+
+## 2.0.78
+
+### Patch Changes
+
+- ai@6.0.73
+
+## 2.0.77
+
+### Patch Changes
+
+- ai@6.0.72
+
 ## 2.0.76
 
 ### Patch Changes