+++ packages/mcp/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ai-sdk/mcp
 
+## 1.0.19
+
+### Patch Changes
+
+- Updated dependencies [7168375]
+  - @ai-sdk/provider@3.0.8
+  - @ai-sdk/provider-utils@4.0.14
+
 ## 1.0.18
 
 ### Patch Changes