+++ packages/rsc/tests/e2e/next-server/CHANGELOG.md
@@ -4,6 +4,32 @@
 
 ### Patch Changes
 
+- Updated dependencies [7168375]
+  - ai@6.0.75
+
+## 0.0.1
+
+### Patch Changes
+
+- Updated dependencies [471009b]
+  - ai@6.0.74
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.73
+
+## 0.0.1
+
+### Patch Changes
+
+- ai@6.0.72
+
+## 0.0.1
+
+### Patch Changes
+
 - ai@6.0.71
 
 ## 0.0.1